+++ packages/@headlessui-react/src/components/focus-trap/focus-trap.tsx
@@ -25,6 +25,26 @@ import { microTask } from '../../utils/micro-task'
 import { useWatch } from '../../hooks/use-watch'
 import { useDisposables } from '../../hooks/use-disposables'
 
+type Containers =
+  // Lazy resolved containers
+  | (() => Iterable<HTMLElement>)
+
+  // List of containers
+  | MutableRefObject<Set<MutableRefObject<HTMLElement | null>>>
+
+function resolveContainers(containers?: Containers): Set<HTMLElement> {
+  if (!containers) return new Set<HTMLElement>()
+  if (typeof containers === 'function') return new Set(containers())
+
+  let all = new Set<HTMLElement>()
+  for (let container of containers.current) {
+    if (container.current instanceof HTMLElement) {
+      all.add(container.current)
+    }
+  }
+  return all
+}
+
 let DEFAULT_FOCUS_TRAP_TAG = 'div' as const
 
 enum Features {
@@ -50,7 +70,7 @@ enum Features {
 export type FocusTrapProps<TTag extends ElementType> = Props<TTag> & {
   initialFocus?: MutableRefObject<HTMLElement | null>
   features?: Features
-  containers?: MutableRefObject<Set<MutableRefObject<HTMLElement | null>>>
+  containers?: Containers
 }
 
 function FocusTrapFn<TTag extends ElementType = typeof DEFAULT_FOCUS_TRAP_TAG>(
@@ -109,8 +129,8 @@ function FocusTrapFn<TTag extends ElementType = typeof DEFAULT_FOCUS_TRAP_TAG>(
       }
     },
     onBlur(e: ReactFocusEvent) {
-      let allContainers = new Set(containers?.current)
-      allContainers.add(container)
+      let allContainers = resolveContainers(containers)
+      if (container.current instanceof HTMLElement) allContainers.add(container.current)
 
       let relatedTarget = e.relatedTarget
       if (!(relatedTarget instanceof HTMLElement)) return
@@ -123,7 +143,7 @@ function FocusTrapFn<TTag extends ElementType = typeof DEFAULT_FOCUS_TRAP_TAG>(
       // Blur is triggered due to focus on relatedTarget, and the relatedTarget is not inside any
       // of the dialog containers. In other words, let's move focus back in!
       if (!contains(allContainers, relatedTarget)) {
-        // Was the blur invoke via the keyboard? Redirect to the next in line.
+        // Was the blur invoked via the keyboard? Redirect to the next in line.
         if (recentlyUsedTabKey.current) {
           focusIn(
             container.current as HTMLElement,
@@ -135,7 +155,7 @@ function FocusTrapFn<TTag extends ElementType = typeof DEFAULT_FOCUS_TRAP_TAG>(
           )
         }
 
-        // It was invoke via something else (e.g.: click, programmatically, ...). Redirect to the
+        // It was invoked via something else (e.g.: click, programmatically, ...). Redirect to the
         // previous active item in the FocusTrap
         else if (e.target instanceof HTMLElement) {
           focusElement(e.target)
@@ -308,7 +328,7 @@ function useFocusLock(
   }: {
     ownerDocument: Document | null
     container: MutableRefObject<HTMLElement | null>
-    containers?: MutableRefObject<Set<MutableRefObject<HTMLElement | null>>>
+    containers?: Containers
     previousActiveElement: MutableRefObject<HTMLElement | null>
   },
   enabled: boolean
@@ -323,8 +343,8 @@ function useFocusLock(
       if (!enabled) return
       if (!mounted.current) return
 
-      let allContainers = new Set(containers?.current)
-      allContainers.add(container)
+      let allContainers = resolveContainers(containers)
+      if (container.current instanceof HTMLElement) allContainers.add(container.current)
 
       let previous = previousActiveElement.current
       if (!previous) return
@@ -348,9 +368,9 @@ function useFocusLock(
   )
 }
 
-function contains(containers: Set<MutableRefObject<HTMLElement | null>>, element: HTMLElement) {
+function contains(containers: Set<HTMLElement>, element: HTMLElement) {
   for (let container of containers) {
-    if (container.current?.contains(element)) return true
+    if (container.contains(element)) return true
   }
 
   return false